# gest-ashoex-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Health Check",
        "description": "API para verificar el estado de salud del sistema",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/facilidades": {
            "get": {
                "tags": [
                    "Facilidades"
                ],
                "summary": "Lista todas las facilidades disponibles",
                "description": "Devuelve una lista de todas las facilidades almacenadas en la base de datos.",
                "operationId": "555dbf6d829c4b2df5967d86a0cfdf29",
                "responses": {
                    "200": {
                        "description": "Lista de facilidades recuperada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nombre_facilidad": {
                                                        "type": "string",
                                                        "example": "Televisor"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "error": {
                                            "type": "null",
                                            "example": null
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lista de facilidades recuperada exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error interno del servidor"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Facilidades"
                ],
                "summary": "Almacena una nueva facilidad en la base de datos",
                "operationId": "3c1b8167301ba7d0015e86507aad758e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre_facilidad"
                                ],
                                "properties": {
                                    "nombre_facilidad": {
                                        "type": "string",
                                        "example": "Televisor"
                                    },
                                    "aulas": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Facilidad registrada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "nombre_facilidad": {
                                                    "type": "string",
                                                    "example": "Televisor"
                                                },
                                                "id_facilidad": {
                                                    "type": "integer",
                                                    "example": 10
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "error": {
                                            "type": "null",
                                            "example": null
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Facilidad registrada exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {},
                                            "example": []
                                        },
                                        "error": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "status": {
                                                        "type": "string",
                                                        "example": "422"
                                                    },
                                                    "detail": {
                                                        "type": "string",
                                                        "example": "El campo nombre_facilidad es obligatorio."
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "example": [
                                                {
                                                    "status": 422,
                                                    "detail": "El campo nombre facilidad es obligatorio."
                                                },
                                                {
                                                    "status": 422,
                                                    "detail": "El campo aula seleccionada no existe."
                                                },
                                                {
                                                    "status": 422,
                                                    "detail": "El campo aula es obligatorio."
                                                },
                                                {
                                                    "status": 422,
                                                    "detail": "El valor del campo nombre facilidad ya está en uso."
                                                }
                                            ]
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}